refactor: update module name
