+++ faster_sam/cache/cache_interface.py
